Move all tests to specs
