+++ crates/rome_js_analyze/src/analyzers/nursery/no_excessive_complexity.rs
@@ -622,227 +622,3 @@ fn calculate_for_variable_declaration(
         })
         .sum()
 }
-
-#[cfg(test)]
-mod tests {
-    use super::calculate_cognitive_complexity;
-    use rome_js_parser::{parse_script, JsParserOptions};
-    use rome_js_syntax::AnyJsFunctionBody;
-    use rome_rowan::AstNodeList;
-
-    #[test]
-    fn test_cognitive_complexity_of_simple_branches() {
-        let body = parse_function(
-            "function simpleBranches() {
-                if (firstCondition) {                 // +1
-                    return 1;
-                } else if (secondCondition) {         // +1
-                    return 2;
-                } else {                              // +1
-                    return 3;
-                }
-            }",
-        );
-        assert_eq!(calculate_cognitive_complexity(&body), 3);
-    }
-
-    #[test]
-    fn test_cognitive_complexity_of_boolean_operators() {
-        let body = parse_function(
-            "function booleanOperators() {
-                if (a                                     // +1 for `if`
-                    && b && c                             // +1
-                    || d || e                             // +1
-                    && f) {                               // +1
-                    return true;
-                }
-            }",
-        );
-        assert_eq!(calculate_cognitive_complexity(&body), 4);
-
-        let body = parse_function(
-            "function booleanOperators2() {
-                if (a                                     // +1 for `if`
-                    &&                                    // +1
-                    !(b && c)) {                          // +1
-                    return true;
-                }
-            }",
-        );
-        assert_eq!(calculate_cognitive_complexity(&body), 3);
-    }
-
-    #[test]
-    fn test_cognitive_complexity_of_sum_of_primes() {
-        let body = parse_function(
-            "function sumOfPrimes(max) {
-                let total = 0;
-                OUT: for (let i = 1; i <= max; ++i) { // +1
-                    for (let j = 2; j < i; ++j) {     // +2
-                        if (i % j == 0) {             // +3
-                            continue OUT;             // +1
-                        }
-                    }
-                    total += 1;
-                }
-                return total;
-            }",
-        );
-        assert_eq!(calculate_cognitive_complexity(&body), 7);
-    }
-
-    #[test]
-    fn test_cognitive_complexity_of_get_words() {
-        let body = parse_function(
-            r#"function getWords(num) {
-                switch (num) {                        // +1
-                    case 1:
-                        return "one";
-                    case 2:
-                        return "a couple";
-                    case 3:
-                        return "a few";
-                    default:
-                        return "lots";
-                }
-            }"#,
-        );
-        assert_eq!(calculate_cognitive_complexity(&body), 1);
-    }
-
-    #[test]
-    fn test_cognitive_complexity_of_nested_flow_structures() {
-        let body = parse_function(
-            r#"function myMethod(num) {
-                try {
-                    if (condition1) {                        // +1
-                        for (let i = 0; i < 10; i++) {       // +2 (nesting=1)
-                            while (condition2) { /* ... */}  // +3 (nesting=2)
-                        }
-                    }
-                } catch (error) {                            // +1
-                    if (condition2) { /* ... */}             // +2 (nesting=1)
-                }
-            }"#,
-        );
-        assert_eq!(calculate_cognitive_complexity(&body), 9);
-    }
-
-    #[test]
-    fn test_cognitive_complexity_of_complex_event_handler() {
-        let body = parse_function(
-            r#"function handleArrowDown(event: React.KeyboardEvent) {
-                const state = getState();
-                const focusedCell = selectFocusedCellOrSurrogate(state);
-                if (!focusedCell) { // +1 = 1
-                    return;
-                }
-              
-                const extendSelection = event.shiftKey;
-              
-                let coordinates: CursorCoordinates | null = null;
-                const focus = selectNotebookFocus(state);
-                const field = getField(focus);
-                const containerEl = getContainerElForCellField(focusedCell.id, field);
-                const containerRect = containerEl && containerEl.getBoundingClientRect(); // +1 = 2
-                if (containerEl && focus.type !== "none" && !event.altKey) { // +2 = 4
-                    const text = selectCellText(state, focusedCell, field);
-                    coordinates = getCoordinatesForOffset(containerEl, text, getFocusOffset(focus));
-              
-                    const lineHeight = getLineHeightForContainer(containerEl);
-                    if (coordinates && coordinates.y + lineHeight < containerRect!.height) { // +3 = 7
-                        // Move the cursor within the cell if we can:
-                        const offset = getOffsetForCoordinates(containerEl, text, {
-                            x: coordinates.x,
-                            y: coordinates.y + lineHeight,
-                        });
-                        dispatch(
-                            focusCell({ cellId: focusedCell.id, field, offset, extendSelection })
-                        );
-                        return true;
-                    }
-                }
-              
-                if (!event.altKey) { // +1 = 8
-                    const targetField = selectRelativeField(state, focusedCell.id, field, 1);
-                    if (targetField) { // +2 = 10
-                        const text = selectCellText(
-                            state,
-                            selectCellOrSurrogate(state, targetField.cellId)!,
-                            targetField.field
-                        );
-                        dispatch(
-                            focusCell({
-                                cellId: targetField.cellId,
-                                field: targetField.field,
-                                offset: charCount(text),
-                                extendSelection: false,
-                            })
-                        );
-                        return true;
-                    }
-                }
-              
-                const targetCell = selectRelativeCellOrSurrogate(state, focusedCell.id, 1);
-                if (!targetCell) {  // +1 = 11
-                    return handleEnd(event);
-                }
-              
-                if (event.altKey) { // +1 = 12
-                    if (focusedCell.readOnly) { // +2 = 14
-                        CellById.get(focusedCell.id)?.shake();
-                        return;
-                    }
-              
-                    if (isSurrogateId(targetCell.id)) { // +2 = 16
-                        // TODO: Should we nudge?
-                    } else { // +1 = 17
-                        // Swap cells with Alt modifier:
-                        dispatch(swapCells(focusedCell.id, targetCell.id));
-                    }
-                } else if (isContentCell(targetCell)) { // +1 = 18
-                    // Move to the cell above and try to maintain the cursor position:
-                    const field = undefined;
-                    const containerBelowEl = getContainerElForCellField(targetCell.id, field);
-                    const extendSelection = event.shiftKey;
-                    if (containerRect && containerBelowEl) { // +3 = 21
-                        const deltaX =
-                            containerRect.left - containerBelowEl.getBoundingClientRect().left;
-                        const lineHeight = getLineHeightForContainer(containerEl);
-                        const offset = coordinates // +3 = 24
-                            ? getOffsetForCoordinates(containerBelowEl, targetCell.content, {
-                                x: coordinates.x + deltaX,
-                                y: lineHeight / 2,
-                            })
-                            : 0;
-                        dispatch(focusCell({ cellId: targetCell.id, offset, extendSelection }));
-                    } else { // +1 = 25
-                        dispatch(focusCell({ cellId: targetCell.id, offset: 0, extendSelection }));
-                    }
-                } else { // +1 = 26
-                    // Move to a cell without cursor position:
-                    dispatch(focusCell({ cellId: targetCell.id, extendSelection }));
-                }
-              
-                return true;
-            }"#,
-        );
-        let complexity_calculated_by_sonarjs = 26;
-        assert_eq!(
-            calculate_cognitive_complexity(&body),
-            complexity_calculated_by_sonarjs
-        );
-    }
-
-    fn parse_function(text: &str) -> AnyJsFunctionBody {
-        let node = parse_script(text, JsParserOptions::default())
-            .tree()
-            .statements()
-            .first()
-            .unwrap();
-        let declaration = node.as_js_function_declaration().unwrap();
-        let body = declaration.body().unwrap();
-
-        AnyJsFunctionBody::JsFunctionBody(body)
-    }
-}
