+++ crates/rome_js_analyze/tests/specs/nursery/noExcessiveComplexity/booleanOperators.js
@@ -1,0 +1,8 @@
+function booleanOperators() {
+    if (a                                     // +1 for `if`
+        && b && c                             // +1
+        || d || e                             // +1
+        && f) {                               // +1
+        return true;
+    }
+}
