+++ crates/rome_js_analyze/tests/specs/nursery/noExcessiveComplexity/booleanOperators2.js
@@ -1,0 +1,7 @@
+function booleanOperators2() {
+    if (a                                     // +1 for `if`
+        &&                                    // +1
+        !(b && c)) {                          // +1
+        return true;
+    }
+}
