+++ crates/rome_js_analyze/tests/specs/nursery/noExcessiveComplexity/booleanOperators2.options.json
@@ -1,0 +1,15 @@
+{
+	"$schema": "../../../../../../npm/rome/configuration_schema.json",
+	"linter": {
+		"rules": {
+			"nursery": {
+				"noExcessiveComplexity": {
+					"level": "error",
+					"options": {
+						"maxAllowedComplexity": 2
+					}
+				}
+			}
+		}
+	}
+}
