+++ crates/rome_js_analyze/tests/specs/nursery/noExcessiveComplexity/getWords.options.json
@@ -1,0 +1,15 @@
+{
+	"$schema": "../../../../../../npm/rome/configuration_schema.json",
+	"linter": {
+		"rules": {
+			"nursery": {
+				"noExcessiveComplexity": {
+					"level": "error",
+					"options": {
+						"maxAllowedComplexity": 1
+					}
+				}
+			}
+		}
+	}
+}
